requirements.txt
pyserial ==3.5
tinydb == 4.8.0
pandas == 2.2.2
openpyxl == 3.1.2
auto-py-to-exe
PyQt5Designer==5.14.1
PyQt5==5.15.9
pyqt5-tools~=5.15
PyQt5-stubs
mysqlclient==2.2.4